pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "openfront-notifier"
version = "0.1.0"
requires-python = ">=3.9"
description = "Notifies you when there is an Openfront game about to start that matches your desired filters."
dependencies = [
    "websockets",
    "websocket-client"
]

[project.scripts]
openfront = "app.cli:main"